requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
httpx==0.27.0
anthropic>=0.40.0
python-dotenv==1.0.1
pydantic>=2.0.0
